Clean up redundant mp4 code and a few other small issues.
