+++ camera_hub/src/lib.rs
@@ -17,7 +17,6 @@ cfg_if! {
         mod version;
         mod notification_target;
         mod mp4;
-        mod fmp4;
         mod android;
         use crate::android::android_dual_stream::get_available_specs;
         use crate::core::{
