+++ camera_hub/src/mp4/mod.rs
@@ -1,0 +1,7 @@
+//! Mp4 camera backend.
+//!
+//! SPDX-License-Identifier: GPL-3.0-or-later
+
+pub(crate) mod mp4_camera;
+pub(crate) mod mp4;
+pub(crate) mod fmp4;
